[diffusion] refactor: rename realtime timer module (#27699)
